+++ src/Ractive/static/adaptors/array/patch.js
