+++ src/Ractive/static/adaptors/array/processWrapper.js
