+++ src/Ractive/static/adaptors/magic.js
